fix split image nodata bug

+++ benchmark/aux/preprocess.py
@@ -2,6 +2,8 @@ import os
 from os.path import basename, join
 from subprocess import call
 
+import numpy as np
+
 import rasterio
 import rastervision as rv
 from rasterio.windows import Window
@@ -25,33 +27,36 @@ def split_image(image_uri, split_dir):
                 continue
             wins.append(Window(c, r, win_size, win_size))
 
-    for i, win in enumerate(wins):
+    i = 0
+    for win in wins:
         with rasterio.open(image_uri) as src:
             img = src.read(window=win)
             win_transform = src.window_transform(win)
             kwargs = src.meta.copy()
 
-        area = image_uri.split('/')[-3]
-        image_id = image_uri.split('/')[-2]
-        output_uri = join(split_dir, area, image_id,
-                          '{}_{}.tif'.format(image_id, i))
-        kwargs.update({
-            'height': win_size,
-            'width': win_size,
-            'transform': win_transform
-        })
+        if np.max(img[-1]) == 255:
+            area = image_uri.split('/')[-3]
+            image_id = image_uri.split('/')[-2]
+            output_uri = join(split_dir, area, image_id,
+                            '{}_{}.tif'.format(image_id, i))
+            kwargs.update({
+                'height': win_size,
+                'width': win_size,
+                'transform': win_transform
+            })
 
-        tmp_uri = join('/tmp/', basename(output_uri))
-        with rasterio.open(tmp_uri, 'w', **kwargs) as dst:
-            dst.write(img)
-        tmp_cmpr_file = tmp_uri.replace('.tif', '_jpg.tif')
-        gdal_command = 'gdal_translate {} {} '.format(tmp_uri, tmp_cmpr_file) +\
-            '-co COMPRESS=JPEG -co JPEG_QUALITY=100 -co TILED=YES ' +\
-            '-co COPY_SRC_OVERVIEWS=YES -co BLOCKXSIZE=512 -co BLOCKYSIZE=512 --config COMPRESS_OVERVIEW JPEG'
-        call(gdal_command, shell=True)
-        upload_or_copy(tmp_cmpr_file, output_uri)
-        for t in (tmp_uri, tmp_cmpr_file):
-            os.remove(t)
+            tmp_uri = join('/tmp/', basename(output_uri))
+            with rasterio.open(tmp_uri, 'w', **kwargs) as dst:
+                dst.write(img)
+            tmp_cmpr_file = tmp_uri.replace('.tif', '_jpg.tif')
+            gdal_command = 'gdal_translate {} {} '.format(tmp_uri, tmp_cmpr_file) +\
+                '-co COMPRESS=JPEG -co JPEG_QUALITY=100 -co TILED=YES ' +\
+                '-co COPY_SRC_OVERVIEWS=YES -co BLOCKXSIZE=512 -co BLOCKYSIZE=512 --config COMPRESS_OVERVIEW JPEG'
+            call(gdal_command, shell=True)
+            upload_or_copy(tmp_cmpr_file, output_uri)
+            for t in (tmp_uri, tmp_cmpr_file):
+                os.remove(t)
+            i += 1
 
 
 class PreProcessCommand(rv.AuxCommand):
